Finish work on refactoring the Amqp message property intercepter and updated tests to mock out the needed bits.

+++ qpid-jms-client/src/test/java/org/apache/qpid/jms/provider/amqp/message/AmqpJmsMessageFacadeTest.java
@@ -193,7 +193,7 @@ public class AmqpJmsMessageFacadeTest extends AmqpJmsMessageTypesTestCase  {
         int producerTtl = 10;
 
         AmqpJmsMessageFacade amqpMessageFacade = createReceivedMessageFacade(createMockAmqpConsumer(), message);
-        amqpMessageFacade.setAmqpTimeToLiveOverride(overrideTtl);
+        amqpMessageFacade.setAmqpTimeToLiveOverride((long) overrideTtl);
 
         amqpMessageFacade.onSend(false, false, producerTtl);
